requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.7
pydantic[email]==2.6.1
python-dotenv==1.0.1
websockets==12.0
pytz==2024.1
PyJWT==2.8.0
httpx==0.24.0
pytest==7.4.0
pytest-asyncio==0.21.0